move const file for import. add font awesome prefixes for set intersection
